pyproject.toml
[project]
name = "gfn-pipeline"
version = "0.1.0"
description = "Global Footprint Network carbon footprint ingestion pipeline (local + AWS + Snowflake)"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "pandas>=2.1.0",
  "pyarrow>=14.0.0",
  "duckdb>=0.10.0",
  "boto3>=1.34.0",
  "tenacity>=8.2.3",
  "python-dotenv>=1.0.1",
  "snowflake-connector-python[pandas]>=3.6.0",
  "dlt[duckdb,snowflake,filesystem]>=1.0.0",
  "aiohttp>=3.9.0",
  "orjson>=3.9.0",
]

[tool.uv]
# uv will manage the isolated environment from this project metadata.

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
